+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Add Kura::Client#cancel_job. Support jobs.cancel API.
   https://cloud.google.com/bigquery/docs/reference/v2/jobs/cancel
+* Kura::Client#wait_job accept Google::Apis::BigqueryV2::Job instance.
 
 # 0.2.1